requirements.txt
acora==2.0
alabaster==0.7.9
anaconda-client==1.6.0
anaconda-navigator==1.5
anaconda-project==0.4.1
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
args==0.1.0
asn1crypto==0.24.0
astroid==1.4.9
astropy==1.3
autopep8==1.3.3
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bioc==1.1.dev2
biopython==1.68
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.4
boto==2.48.0
Bottleneck==1.2.0
brewer2mpl==1.4.1
bz2file==0.98
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
chembl-webresource-client==0.9.23
chest==0.2.3
CIRpy==1.0.2
click==6.7
clint==0.5.1
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
conda==4.4.10
configobj==5.0.6
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
cymem==1.31.2
Cython==0.25.2
cytoolz==0.9.0
dask==0.13.0
datashape==0.5.4
decorator==4.1.2
demjson==2.2.4
dill==0.2.5
dlib==19.9.0
docutils==0.13.1
easydict==1.7
en-core-web-sm==1.2.0
et-xmlfile==1.0.1
ete3==3.0.0b35
face-recognition==1.2.1
face-recognition-models==0.3.0
fake-useragent==0.1.7
fastcache==1.0.2
feather-format==0.4.0
Flask==0.12
Flask-Cors==3.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
floyd-cli==0.10.21
ftfy==4.4.3
future==0.16.0
geniatagger-python==0.3
gensim==2.3.0
gevent==1.2.2
gevent-openssl==1.2
ggplot==0.11.5
glove-python==0.1.0
greenlet==0.4.11
grequests==0.2.0
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
intervaltree==2.1.0
ipykernel==4.5.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.11.0
Jinja2==2.10
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.3.3
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.1
jupyter-nbextensions-configurator==0.2.8
jupyterthemes==0.18.2
Keras==2.1.1
lazy-object-proxy==1.2.2
lesscpy==0.12.0
llvmlite==0.15.0
locket==0.2.0
lxml==4.1.1
marisa-trie==0.7.4
MarkupSafe==1.0
marshmallow==2.14.0
matplotlib==2.1.0
missingno==0.3.8
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.9
murmurhash==0.26.4
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.2
NoAho==0.9.6.1
nose==1.3.7
notebook==4.3.1
numba==0.30.1
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.6.0
odo==0.5.0
olefile==0.45.1
opencv-python==3.4.0.12
openpyxl==2.4.1
packaging==16.8
pandas==0.22.0
parso==0.1.0
partd==0.3.7
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.0.0
plac==0.9.6
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.2.0
psutil==5.0.1
ptyprocess==0.5.2
PubChemPy==1.0.4
pubmed-parser==0.1
py==1.4.34
pyahocorasick==1.1.4
pyamg==3.3.0
pyarrow==0.8.0
pyasn1==0.4.2
pycodestyle==2.3.1
pycorenlp==0.3.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pymetamap==0.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyPdf==1.13
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
raven==6.3.0
redis==2.10.5
regex==2017.7.28
requests==2.18.4
requests-cache==0.4.13
requests-toolbelt==0.8.0
rope-py3k==0.9.4.post1
ruamel-yaml==0.15.35
scikit-image==0.12.3
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.7.1
selenium==3.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sortedcontainers==1.5.7
spacy==1.9.0
spherecluster==0.1.5
Sphinx==1.5.1
splitta==0.1.0
spyder==3.1.2
SQLAlchemy==1.1.5
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tabulate==0.8.1
tensorflow==1.0.1
termcolor==1.1.0
terminado==0.6
Theano==1.0.0
thinc==6.5.2
toolz==0.9.0
tornado==4.4.2
tqdm==4.15.0
traitlets==4.3.2
ujson==1.35
unicodecsv==0.14.1
Unidecode==0.4.21
unittest2py3k==0.5.1
unittest2six==0.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.15
widgetsnbextension==1.2.6
wrapt==1.10.8
WTForms==2.1
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
yapf==0.20.0
